add some error checking.

+++ sabnzbd_upgrade.sh
@@ -32,48 +32,52 @@ START_UPDATE="0"
 #
 ############################################################################################################
 
-LOCAL_VERSION=`cat ${SABDESTDIR}/${SABDIR}/PKG-INFO | grep ^Version | awk '{print $2}'`
-VERSION=`curl -s http://sabnzbdplus.sourceforge.net/version/latest | head -n1`
+if [ -f ${SABDESTDIR}/${SABDIR}/PKG-INFO ] && [ -f ${SABCONFIGDIR}/sabnzbd.ini ]; then
+	LOCAL_VERSION=`cat ${SABDESTDIR}/${SABDIR}/PKG-INFO | grep ^Version | awk '{print $2}'`
+	VERSION=`curl -s http://sabnzbdplus.sourceforge.net/version/latest | head -n1`
 
-if [ "${VERSION}" != "${LOCAL_VERSION}" ]; then
-START_UPDATE="1"
-fi
-
-if [ "${START_UPDATE}" = "0" ]; then
-        echo "You are on the current version of: ${LOCAL_VERSION}"
-        echo "No update will be performed"
-else
-	cd ${SABDESTDIR}
-	echo "New version found! You are on ${LOCAL_VERSION}, upgrading to ${VERSION} now."
-
-	API_KEY=`cat ${SABCONFIGDIR}/sabnzbd.ini | grep ^api_key | awk '{print $3}'`
-	PORT=`cat ${SABCONFIGDIR}/sabnzbd.ini | grep ^port | awk '{print $3}' | head -1`
-	DIR="SABnzbd-${VERSION}"
-	GZ="${DIR}-src.tar.gz"
-
-	echo "Downloading SABnzbd ${VERSION} (${GZ})"
-	curl -s -C - -O "${SABURL}/${VERSION}/${GZ}"
-
-	echo "Unpacking ${GZ}"
-	tar -xzf ${GZ} && rm ${GZ}
-
-	echo "Shutting down SABnzbd+"
-	curl -s "http://${HOST}:${PORT}/sabnzbd/api?mode=shutdown&apikey=${API_KEY}" >> /dev/null
-
-	if [ "${ARCHIVE}" = "0" ]; then
-		rm -rf ${SABDIR}
-	else
-		echo "Archiving old SABnzbd+"
-		mkdir -p ${ARCHIVEDIR}
-		mv ${SABDIR} ${ARCHIVEDIR}/${SABDIR}_${LOCAL_VERSION}_`date +'%Y%m%d-%H%M'`
+	if [ "${VERSION}" != "${LOCAL_VERSION}" ]; then
+		START_UPDATE="1"
 	fi
-	echo "Installing new SABnzbd+"
-	mv ${DIR} ${SABDIR}
 
-	echo "Restarting SABnzdb+"
-	python ${SABDIR}/SABnzbd.py -d -f ${SABCONFIGDIR}/sabnzbd.ini >> /dev/null
+	if [ "${START_UPDATE}" = "0" ]; then
+        	echo "You are on the current version of: ${LOCAL_VERSION}"
+	        echo "No update will be performed"
+	else
+		cd ${SABDESTDIR}
+		echo "New version found! You are on ${LOCAL_VERSION}, upgrading to ${VERSION} now."
 
-	# Go back to the previous directory
-	cd -
-	echo "Upgrade complete !"
+		API_KEY=`cat ${SABCONFIGDIR}/sabnzbd.ini | grep ^api_key | awk '{print $3}'`
+		PORT=`cat ${SABCONFIGDIR}/sabnzbd.ini | grep ^port | awk '{print $3}' | head -1`
+		DIR="SABnzbd-${VERSION}"
+		GZ="${DIR}-src.tar.gz"
+
+		echo "Downloading SABnzbd ${VERSION} (${GZ})"
+		curl -s -C - -O "${SABURL}/${VERSION}/${GZ}"
+
+		echo "Unpacking ${GZ}"
+		tar -xzf ${GZ} && rm ${GZ}
+
+		echo "Shutting down SABnzbd+"
+		curl -s "http://${HOST}:${PORT}/sabnzbd/api?mode=shutdown&apikey=${API_KEY}" >> /dev/null
+
+		if [ "${ARCHIVE}" = "0" ]; then
+			rm -rf ${SABDIR}
+		else
+			echo "Archiving old SABnzbd+"
+			mkdir -p ${ARCHIVEDIR}
+			mv ${SABDIR} ${ARCHIVEDIR}/${SABDIR}_${LOCAL_VERSION}_`date +'%Y%m%d-%H%M'`
+		fi
+		echo "Installing new SABnzbd+"
+		mv ${DIR} ${SABDIR}
+
+		echo "Restarting SABnzdb+"
+		python ${SABDIR}/SABnzbd.py -d -f ${SABCONFIGDIR}/sabnzbd.ini >> /dev/null
+
+		# Go back to the previous directory
+		cd -
+		echo "Upgrade complete !"
+	fi
+else
+	echo "There is an error with the variables chosen.  Please edit the script and try again."
 fi
